remove transform alias too
